refactor(api): move certification-point-of-contact serializer
in identity-access-management
