+++ api/src/identity-access-management/infrastructure/serializers/jsonapi/certification-point-of-contact.serializer.js
